pyproject.toml
[tool.poetry]
name = "etl-test"
version = "0.1.0"
description = ""
authors = ["carigonz <carolinagonzalez@gmail.com>"]
readme = "README.md"
packages = [{include = "airflow-etl-project"}]

[tool.poetry.dependencies]
python = "^3.9, <3.13"
pandas = "^1.5.3"
pdfplumber = "^0.11.4"
pypdf2 = "^3.0.1"
gspread = "^6.1.3"
oauth2client = "^4.1.3"
opencv-python = "^4.10.0.84"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.149.0"
requests = "^2.32.3"
numpy = "^1.26.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
sqlalchemy = "1.4.36"
sqlalchemy-redshift = "0.8.14"
apache-airflow = "^2.10.2"
apache-airflow-providers-openlineage = "^1.8.0"
pyarrow = "^14.0.1"
snowflake-connector-python = "^3.5.0"
apache-airflow-providers-google = "^10.11.0"
apache-airflow-providers-snowflake = "^5.1.1"
pybind11 = "^2.12"
poetry-plugin-export = "^1.8.0"
flake8 = "^5.0.0"
autopep8 = "^2.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
